Expose standars field configs from grafana ui via function
